Fix warning

+++ src/core/TransfersManager.cpp
@@ -192,7 +192,7 @@ void Transfer::timerEvent(QTimerEvent *event)
 
 				speedSum /= m_speeds.count();
 
-				m_remainingTime = qRound(static_cast<qreal>(m_bytesTotal - m_bytesReceived) / speedSum);
+				m_remainingTime = qRound(static_cast<qreal>(m_bytesTotal - m_bytesReceived) / static_cast<qreal>(speedSum));
 			}
 
 			emit changed();
